==> scripts/2_zalando_dataset_get_images.ipynb <==
import os
import pickle
import pandas as pd
import requests
import random
import time
from fake_useragent import UserAgent

with open('items.pkl', 'rb') as file:
    items = pickle.load(file)

print(type(items), len(items))

first_key = next(iter(items.keys()))
first_key

print(items[first_key])

if os.path.exists('downloaded_items.pkl'):
    with open('downloaded_items.pkl', 'rb') as file:
        saved_items = pickle.load(file)
else:
    saved_items = {
    'damenbekleidung-kleider': [],
    'damenbekleidung-shirts': [],
    'damenbekleidung-jeans': [],
    'damenbekleidung-jacken': [],
    'damenbekleidung-jacken-maentel': [],
    'damenbekleidung-blusen-tuniken': []
}

def download_image(url, save_path):
    ua = UserAgent()
    headers = {"User-Agent": ua.random}

    try:
        if os.path.exists(save_path):
            return 
        os.makedirs(os.path.dirname(save_path), exist_ok=True)
        
        response = requests.get(url, headers=headers, timeout=10)
        if response.status_code == 200:
            with open(save_path, 'wb') as file:
                file.write(response.content)
        else:
            print(f"Failed to download {url} - HTTP {response.status_code}")
    except Exception as e:
        with open('downloaded_items.pkl', 'wb') as file:
             pickle.dump(saved_items, file)
            
        print(f"Error downloading {url}: {e}")

MAX_ITEMS = 5300
count = 0
for item in items.values():
    if item['id'] in saved_items[item['category']]:
        continue
    if len(saved_items[item['category']]) < MAX_ITEMS:
        for i in range(len(item['images'])):
            save_path = 'images/' + item['images'][i].replace('https://img01.ztat.net/article/', '')
            download_image(item['images'][i], save_path)
            if count % 100 == 0:  # Rest after every 100 requests
                print(f"Downloaded: {count} items.")
                time.sleep(random.uniform(5, 10))
            else:
                time.sleep(random.uniform(1, 2))  # Shorter random delay
            count += 1
        
        saved_items[item['category']].append(item['id'])
        print(f"Downloaded: {item['category']}, {item['id']}")

with open('downloaded_items.pkl', 'wb') as file:
    pickle.dump(saved_items, file)

for item in saved_items.keys():
    print(f"Downloaded: {len(saved_items[item])} {item}")
    
#saved_items
